feat: add helper for BP entry element
